College of Medicine - Effort rounding error (#60)
* Convert variable type for "hours" to decimal type

* event_date years are input as 4 digits not 2

+++ app/importers/com_data/com_effort_populate_db.rb
@@ -24,7 +24,7 @@ class ComData::ComEffortPopulateDb
                                                      faculty:)
 
         com_effort.hours ||= 0
-        com_effort.hours += row['UME_CALCULATED_TEACHING_WHILE_NON_BILLING_EFFORT__HRS_'].to_f
+        com_effort.hours += row['UME_CALCULATED_TEACHING_WHILE_NON_BILLING_EFFORT__HRS_'].to_d
 
         com_effort.save!
       rescue ActiveRecord::RecordNotUnique
